+++ src/components/waybill/BillCreate.vue
@@ -475,12 +475,27 @@
       },
 
       deleteByIds(data){
-        this.deleteRequest("/carriers/delete/"+data).then(res=>{
-            if (res){
-              this.loadCarrierss();
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          this.deleteRequest("/carriers/delete/"+data).then(res=>{
+              if (res){
+                this.loadCarrierss();
+              }
             }
-          }
-        )
+          );
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          });
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          });
+        });
       },
       showDialog(data){
         this.dialogFormVisible=true;
